requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
pillow>=10.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.35.0
torch>=2.0.0
pdf2image>=1.16.3
PyPDF2>=3.0.1
numpy>=1.24.0
scipy>=1.11.0